requirements.txt
amqp==5.3.1
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.2.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.7
django-allauth==65.4.1
django-apscheduler==0.7.0
django-celery-beat==2.7.0
django-filter==25.1
django-timezone-field==7.1
idna==3.10
jwt==1.3.1
kombu==5.4.2
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.3.0b5
requests==2.32.3
setuptools==76.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13

django-modeltranslation~=0.19.12